+++ src/api/v2/account.rs
@@ -167,8 +167,6 @@ mod tests {
 
   use time_util::parse_system_time_from_str;
 
-  use url::Url;
-
   use uuid::Uuid;
 
   use crate::api::API_BASE_URL;
@@ -242,12 +240,7 @@ mod tests {
 
   #[test(tokio::test)]
   async fn request_account_with_invalid_credentials() {
-    let api_base = Url::parse(API_BASE_URL).unwrap();
-    let api_info = ApiInfo {
-      base_url: api_base,
-      key_id: "invalid".to_string(),
-      secret: "invalid-too".to_string(),
-    };
+    let api_info = ApiInfo::from_parts(API_BASE_URL, "invalid", "invalid-too").unwrap();
     let client = Client::new(api_info);
     let result = client.issue::<Get>(()).await;