+++ src/api_info.rs
@@ -1,4 +1,4 @@
-// Copyright (C) 2019 Daniel Mueller <deso@posteo.net>
+// Copyright (C) 2019-2021 Daniel Mueller <deso@posteo.net>
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 use std::env::var_os;
@@ -30,6 +30,23 @@ pub struct ApiInfo {
 }
 
 impl ApiInfo {
+  /// Create an `ApiInfo` from the required data.
+  ///
+  /// # Errors
+  /// - [`Error::Url`](crate::Error::Url) If `base_url` cannot be parsed
+  ///   into a [`url::Url`](url::Url).
+  pub fn from_parts(
+    base_url: impl AsRef<str>,
+    key_id: impl ToString,
+    secret: impl ToString,
+  ) -> Result<Self, Error> {
+    Ok(Self {
+      base_url: Url::parse(base_url.as_ref())?,
+      key_id: key_id.to_string(),
+      secret: secret.to_string(),
+    })
+  }
+
   /// Create an `ApiInfo` object with information from the environment.
   ///
   /// This constructor retrieves API related information from the