+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+- Added `ApiInfo::from_parts` constructor
 - Bumped minimum supported Rust version to `1.46`